[ios][calendar] refactor files into groups
