+++ packages/expo-calendar/ios/Records/CalendarRecords.swift
@@ -1,7 +1,8 @@
 import ExpoModulesCore
 
 enum CalendarEntity: String, Enumerable {
-  case event, reminder
+  case event
+  case reminder
 }
 
 struct CalendarRecord: Record {
@@ -75,35 +76,6 @@ struct Event: Record {
   var availability: String
   @Field
   var status: String
-}
-
-struct Alarm: Record {
-  @Field
-  var absoluteDate: String?
-  @Field
-  var relativeOffset: Int?
-  @Field
-  var structuredLocation: AlarmLocation?
-  @Field
-  var method: String?
-}
-
-struct AlarmLocation: Record {
-  @Field
-  var title: String
-  @Field
-  var proximity: String?
-  @Field
-  var radius: Double?
-  @Field
-  var coords: Coordinates?
-}
-
-struct Coordinates: Record {
-  @Field
-  var latitude: Double
-  @Field
-  var longitude: Double
 }
 
 struct RecurringEventOptions: Record {
